Fixes benchmark and some fixes to test
This includes (I hope) the contents of https://github.com/edgexfoundry/edgex-go/pull/302.

More importantly, it addresses an issue in the benchmark that's manifestated twice - in events and in readings. The original behavior was to test AddReading (and AddEvent) to establish the measurement, and then continue with the rest of the benchmarks. However, the faster the database is actually doing the adds, more readings (or events) are created and passed to subsequent benchmarks. Then, doing a full fetch of all objects as in Readings (or Events) will take considerably longer for a faster database (IO capped) than a slower one.

The proposed solution is scrubbing the db after benching the addition, and populating it with a predetermined known static sized (e.g. 1K) dataset.

Signed-off-by: Itamar Haber <itamar@redislabs.com>