snap/snapcraft.yaml: add cassandra db as a service
Another database is necessary for kong to work with the new security services, and options are either postgres or Cassandra, and postgres doesn't work well as a snap because it can't be run as root, so this part/app installs cassandra into the snap by downloading it from the official apache download site. 
Also updated the snap's README to reflect the additional interface that needs to be connected for cassandra to work properly.
The wrapper and such have been modified from : https://github.com/evandandrea/cassandra-snap
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>