snap/bin/stop-edgex.sh: don't let kill or mongod kill the stop script
This is necessary to ensure that when the stop script is called as many of the services are stopped as are able. So if signalling a particular service fails, we shouldn't stop there and try all the other ones.
This is important for example when certain services fail to startup and edgex is in an incomplete state with just consul running and we need to "reset" edgex's services using the stop command. If we failed at say mongo, then we couldn't stop consul using this script and would manually have to kill it which is a bad user experience.
Arguably this could be another script, but it doesn't seem that important when you're stopping the services that you don't kill other services when attempting to kill one thing fails.
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>