snap/snapcraft.yaml: standardize build scripts for go services
This new script makes the process for building a go service or part a standardized process where the part is staged into it's own isolated $GOPATH and it has $GOROOT set as expected using the go install built in the go part, and also has $PATH set to include staging directories so for example glide can easily be used in a consistent manner for all parts.
We also stage glide into $SNAPCRAFT_STAGE/bin, and use $SNAPCRAFT_ARCH_TRIPLET for determining architecture to use.
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>