snap/bin/security-start.sh: add scripting for security services
This adds support for enabling the EdgeX security services within the snap by setting the SECURITY=y variable in the edgex-services-env file. This defaults to y, or on.
This also adds an install hook, which installs various files into $SNAP_DATA and configures some config files to use absolute paths for $SNAP_DATA and things in the config files, as well as generate a unique psuedorandom password for the TLS certificates to be used with the security services.
Lastly, adding support for the stop-edgex script to stop all of the security related processes gracefully.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>