snap/bin/security-start.sh: run initial unseal attempt loop before vault-worker
This ensures that if unsealing the vault fails after 10 attempts, the whole script will end and won't just hang.
Additionally, it allows for a much quicker initial startup as we don't pay the penalty of the watchdog time if we fail the first time.
Lastly, we can reset the watchdog time to 3 minutes as it runs in the background indefinitely, so that doesn't just eat up CPU time trying to continually unseal the vault.
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>