snapcraft.yaml: use config-seed files from edgex-go
Now we use only the config files from edgex-go and remove the core-config-seed-go repo part entirely. 
Also note that this removes the banner.txt file from the snap.
Lastly, we change the location of the config-seed config files to be just $SNAP/config/config-seed, and use the name of the binary as just config-seed
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>