snap/build.sh: delete the login file and image after build
The snap store login file contains information enough for an attacker to publish any possible binary to the store under the edgexfoundry name, and essentially full control over any snaps that edgexfoundry controls.
To seed the docker image with the login file, we have to copy it into the git root for it to be a part of the build context, so we need to delete this file from the git root, in case the build environment for this build is saved/shared at all.
Additionally, we need to delete the image and prune after running the build to ensure that there are not leftover intermediate containers left in the local docker image registry that another job that doesn't have access to these credentials could revive to gain access to the credentials through the build context constructed for this image.
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>