snapcraft.yaml: process config file dirs during install hook
Instead of hard-coding the location of $SNAP_COMMON as /var/snap/edgexfoundry/common, we should use the environment variable. However, because these are config files, they don't get expanded with environment variables during runtime, so instead we put the environment variable in the config file during build time, and replace the value of the variable during the install hook so that this is dynamically set. 
This should allow the snap to be installed and run on Arch for example which doesn't use the standard snap paths as Ubuntu.
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>