start-edgex.sh: go into a safe directory before starting
Currently, when running start-edgex say from one's $HOME directory as sudo, there will be a denial in the syslog about comm="start-config-se" (which I believe is the start-config-seed.sh script) attempting to open $HOME and being denied. EdgeX doesn't need to access this directory so I think this is due to some shell command attempting to read the original working directory, which would be $HOME. We can resolve this by first going into $SNAP_DATA which is a writable location and then running the rest of the scripts from there.
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>