snapcraft.yaml: download jar files for java services
This results in quicker build times and allows easier management of the services' version and repo from the nexus to use when including them in the snap. Ideally we would control this with the `source` spec, but we also need files from the git repo, so this is what we have.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>