snap/many: replace usage of systemd-notify with forking daemons
This change is necessary because the daemon-notify interface is not implemented properly and is missing the capabilities "net_admin" and "sys_admin" in AppArmor in order to notify for a different process than the one that `systemd-notify` runs as. 
To remedy this, we instead make all of the daemons that were using systemd-notify into forking daemons, with an additional forking wrapper which immediately forks a process such that systemd can always immediately think that the process has successfully started up. We also need to ensure that the process that is initially forked stays alive for the remainder of the lifetime of the service because that service will be the "main" PID that systemd tracks. So for the go services we need to use exec instead of just running it normally, and for the security-services wrapper we already are doing the right thing in the script by waiting for all the child processes to exit. This point is doubly important because systemd docs say that forking daemons should specify a PIDFile, which unfortunately we can't do in the snap, as snapd doesn't support this, so if the main PID isn't the PID that systemd sees initially, bad things will happen. As currently implemented however it's not a problem.


Signed-off-by: Ian Johnson <ian.johnson@canonical.com>