#942 Remove MemoryDB Provider
Removes the MemoryDB provider and refactors affected
unit tests to leverage mocks instead of relying on
MemoryDB instances.  A small portion of the MemoryDB
provider remains in internal/export/client.  This is
meant to be a short term solution until this package can
be removed or the dependencies on MemoryDB can be refactored
out.

Signed-off-by: Daniel Harms <jdharms@gmail.com>