Return recent events and readings for Query APIs
Modified query APIs to sort the results based on modified time.

Fix #966.  The queries with a limit were just returning the oldest
results.

This change returns the most recent results for queries with a limit.
The results are sorted by the modified time as this can be more recent
than the created time.

queries affected:
 - /event/device/{deviceId}/{limit}:
 - /event/{start}/{end}/{limit}:
 -  /event/device/{deviceId}/valuedescriptor/{valuedescriptor}/{limit}:
 -  /reading/device/{deviceId}/{limit}:
 -  /reading/name/{name}/device/{device}/{limit}:
 -  /reading/name/{name}/{limit}:
 -  /reading/uomlabel/{uomLabel}/{limit}:
 -  /reading/label/{label}/{limit}:
 -  /reading/type/{type}/{limit}:
 -  /reading/{start}/{end}/{limit}:

Possible side effects
1. Performance and scale.  It is recommended to create an index of the
modified time if performance is an issue.
2. the name of the field is hard coded, so if the name changes, this
code will need to be modified.

Signed-off-by: Jennifer <quyrean@yahoo.com>