snap/hooks/install: use current symlink for $SNAP_DATA
If we use $SNAP_DATA verbatim when configuring the various config files
here, it will work for the initial revision installed, but upon a
refresh, those file paths will be inaccessible to the next revision and
won't point to the next revision's separate files.

Using the current symlink solves this, as the current symlink is updated
upon a refresh.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>