many: rename vault-config.json to vault-config.hcl
The vault config file is actually in the HashiCorp Configuration
Language (hcl) and not in JSON, so it is more appropriate to use the .
hcl file extension for that file.

Also change the $SNAP_DATA in the comment to just SNAP_DATA so it 
doesn't get modified by sed when we go to replace $SNAP_DATA at install
time.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>