snap/hooks/{post,pre}-refresh: add hooks for upgrading old snaps
These new post-refresh and pre-refresh hooks handle upgrading to use
the "current" symlink for absolute paths such as $SNAP_DATA that are
inserted verbatim into configuration files, where they become invalid
after a refresh. They become invalid because currently $SNAP_DATA uses
the revision number of the current snap explicitly rather than the
"current" symlink which points to the revision number in that directory.

Note that we do a best effort here on upgrading the paths and try not to
override any changes that the user may have manually made to these files
by only replacing the specific $SNAP_DATA value with any revision number
with the same value, but using the "current" symlink instead.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>