snap/hooks/pre-refresh: set config items before refreshing
This change ensures that in the future we will be able to identify the
release a snap was on before it was refreshed, to allow easier fixes in
the case something is broken and we need to fix.

Additionally it gives us the option to special case Edinburgh -> XYZ
upgrades by using the "release" config item.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>