Integrate model validation, fix tests
- The encapsulation of validation inside the contracts/models required
some unit tests to be reworked. In particular, unit tests that
previously used literal JSON strings as the basis from which to un-
marshal have been replaced with an actual test instance of the model.
This means that if during the course of a test a model is marshaled to
bytes and then unmarshaled, the internal model validation logic will
fire. This was particularly prevalent in the way the export tests are
implemented.
- The LogLevels constants were moved in core-contracts to the models
package and out of the clients package. Had to address that.
- Reworked the relationship of Get/Put properties on Commands to no
longer be pointers. Checking for nil is a poor way to see if a given
instance is empty, plus its error prone in cases where the check isn't
even done.
- Changes related to the removal of the contracts/models/Service struct.
This struct was only utilized by DeviceService and contained most of
that type's properties. It was removed due to lack of re-use.
- The above cleaned up property references on the Device (e.g. We no
longer have to reference something like device.Service.Service.Id)
- Controller validation that previously called anonymous functions to
validate OperatingState and AdminState values have been removed since
those checks are now part of the models themselves.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>