snap/hooks/configure: set config item for service in handle_svc
This ensures that when we use the handle_svc function internally to turn
on or off dependent services, those settings are persistent. For
example, with the snap settings, in order,
snap set edgexfoundry support-logging=off
snap set edgexfoundry device-random=on

Now support-logging is in an incosistent state where it is turned on and
enabled as far as systemd is concerned because we turned it on when we
requested device-random turn on, however the next time the configure
hook runs, it will be stopped and disabled because the snap config
setting for support-logging was not updated to be on. Setting the config
setting when we modify the systemd state helps keep these things in sync

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>