Return New Notification ID Instead of Slug
Fix #1329

When entities are created in EdgeX the normal practice is to return the
ID of the new record. Previously we were returning the "slug" value
when a notification entry was created, which is something that is provided
by the application creating the notification. Thus it wasn't very
useful. Standardized POST of new notification to return the new ID and
updated the RAML.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>