Transmission resend entries created with 0 CreateDate
Fix #1315

If a notification comes in and transmittal to a subscriber fails, a
retry can occur. In this case, the retry transmission record was being
created with a 0 CreateDate value. This was because dates are managed
at the database provider level and the return value on an ADD operation
is simply the ID of the newly inserted record.  In order to fix this,
I had to add a new function to the DBClient and supporting providers
that would fetch the newly inserted Transmission record so that the
date values could be retained for further opertations.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>