Improve delete event and reading performance
Fix #1286

Relocate the responsibility of deleting events by device and deleting
readings by device to the DBClient rather than the business layer.

Add a new function to the DBClient interface as well as the Mongo and
Redis implementations which abstract the process of deleting readings
and events by device ID.

Update the business logic to leverage the newly added functions for more
effectively deleting events and readings associated with the specified
device. Update tests to reflect the new business logic

The Mongo implementation has gone from O(n) to O(1) where n is the
number of events/readings associated with a Device. However, the Redis
implementation still remains at O(n) for event deletion process as a
whole where n is the number of events/readings associated with the
device. The deletion functionality is split into two processing units.
One which renames the entity to be deleted Redis key, this is
functionaly the same as a delete from the perspective of other clients
as they cannot retrieve the original data. The other process cleans up
the lingering data. The API has been updated to rename the Redis keys
which are to be deleted, asynchronously delete the lingering data, and
respond to the user. Therefore, the API executes faster, albeit O(n) due
to iterating over each element for the rename, from the perspective of
the user, but the deletion process as a whole can execute much slower
due to the need of serialization. The reason is because we must retrieve
the full object that is persisted so that we can remove all the
data(Sorted sets) that reside in the data-store. For example, there
exists sorted sets for time created(reading:created, event:created), and
device(reading:device:{Device}, event:device:{Device}). Given only the
device information we are not be able to remove the data from
Redis(ZREM) without constructing the Redis key which is compromised of
the data which is stored.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>