snap/hooks/pre-refresh: set snap release to fuji
This was missed in the PR since the PR was merged after edinburgh was
branched, but for master now we want fuji as the release.

Also set epoch to 2 for fuji to prevent automatic edinburgh -> fuji
upgrades until we have a migration strategy ready that will use an epoch
of 2* (to read epoch 1 and write epoch 2).

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>