snap/hooks/install: disable sec-api-gateway on arm64
See https://github.com/edgexfoundry/blackbox-testing/issues/185 for more
details, but basically Kong isn't meant to work properly on arm64 until
it is updated to use OpenResty 1.15 so until that point we just always
disable the security-api-gateway services on arm64 in the install hook.

Note that due to another snapd bug where all services are restarted on 
refresh, we also need to implement a workaround in the kong service 
itself because it will be started up on refresh and we still don't want
kong to really run on arm64 even in the event it is erroneously started.
Returning from the kong-daemon.sh script like this will result in 
systemd initially assuming that the process started and proceeding to 
start other dependent services, but shortly thereafter realizing there
are not children processes and thus marking the service as "inactive".

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>