snapcraft.yaml: update registry arg for dev svcs
The registry arg is now a --key=val style option, where the value is
a custom URL spec with the scheme as the registry type. Since the snap
only works with consul for now, the scheme is just "consul".

Note we have to specify the value as an environment variable because
snapd doesn't allow the ":" character in a app's command spec. See
https://bugs.launchpad.net/snapd/+bug/1827392 for the upstream
bug/feature request to snapd.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>