Refactor restGetAddressableByPort to use operators
Fix #1489

Refactor the `restGetAddressableByPort` REST function handler to use the
operator pattern in efforts to increase unit test coverage.

Update the `restGetAddressableByPort` to use the PortExecutor operator
rather than directly using the DBClient. This will allow for easy
mocking during testing.

Add tests for the newly added PortExecutor operator and all of the
necessary layers needed by the operator.

Add tests for the REST handler `restGetAddressableByPort` to cover the
different situations the function is expected to handle.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>