Refactor restGetAddressableByTopic to use operator
Fix #1492

Refactor the restGetAddressableByTopic REST function handler to use the
operator pattern in efforts to increase unit test coverage.

Update the restGetAddressableByTopic to use the TopicExecutor operator
rather than directly using the DBClient. This will allow for easy
mocking during testing.

Add tests for the newly added TopicExecutor operator and all of the
necessary layers needed by the operator.

Add tests for the REST handler restGetAddressableByTopic to cover the
different situations the function is expected to handle.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>