Add validation when update device profile
Fix issue #1494

Add checks before updating device profile and replace direct calls to
the DBClient with the operator pattern to allow for easier testing.

Update the restUpdateDeviceProfile rest handler to check if there are
associated devices or provision watchers before performing an update. If
there are associations for the device profile a Conflict(409) is
returned to the user and the update is not persisted. Update the
successful response to return a 204(No Content) with no body rather than
a "true"

Add new `UpdateDeviceProfileExecutor` which provides functionality for
updating device profiles by using the operator pattern. Also, move all
business logic to the `UpdateDeviceProfileExecutor` for segregation of
responsibilities. The REST handler is responsible for HTTP/REST logic
like serializing the request body, where the
`UpdateDeviceProfileExecutor` is responsible for interacting with the
database and ensuring the device profile which is being updated is in
the correct state before performing the update. For example, no devices
nor provision watchers are associated with the device profile.

Update the `restUpdateDeviceProfile` rest handler to use the new
operator.

Add new error types for the different errors the
`UpdateDeviceProfileExecutor` may encounter during processing. This will
allow for type checking for more insight on the underlying error

Update the `restUpdateDeviceProfile` to handle the different types of
errors the `UpdateDeviceProfileExecutor` may return

Add new `ProfileIdExecutor` for obtaining devices by a profile ID using
the operator pattern.

Update `notifyProfileAssociates` function to use the new operator

Added necessary mocks for new interfaces

Remove dead code which checked for duplicate command names within
DeviceProfile. This is dead code due to the validation logic already
handling this scenario and will result in the marshaling process to
short-circuit and result in a 400(Bad Request) response to the user

Add tests for `restUpdateDeviceProfile` REST handler, `UpdateDeviceProfileExecutor`, and `UpdateDeviceProfileExecutor`

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>