Relocate ValueDescriptor responsibility (#1557)
Fix #1471

Add functionality to the `restAddDeviceProfile` and
`restAddDeviceProfile` REST handlers to not only create/update
DeviceProfiles, but also make the necessary changes to the associated
ValueDescriptors. This functionality was previously expected to be in
the Device SDK. Now, we are expecting the SDK functionality to be
updated to only create/update DeviceProfiles, and not ValueDescriptors
as Core-Metadata will now handle that responsibility. See issue #318 for
more details regarding the removal of this functionality from the SDK.

Add new operators `ValueDescriptorAdder` `ValueDescriptorUpdater` which
handles the responsibility of creating/updating and persisting the
necessary ValueDescriptors given a DeviceResources

Add tests for `ValueDescriptorAdder` operator.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>