feature(query-params): Pass QueryParams through EdgeX to Device Services (#1571)
Fix #1564

This pull request is an example/POC that @lenny-intel and I put together of how #1564 could be accomplished. This is only 1/2 of the equation as the device-sdk would need to be updated to handle the query parameters. The good thing with this is that it wouldn't be a breaking change as the device-sdk would simply ignore the query parameters that are sent on.

Signed-off-by: Mike Johanson <michael.johanson@intel.com>