bin/test-attribution-txt.sh: add test for attribution files
Check that all of the modules listed by go mod in the vendor directory when
running `go mod vendor` are attributed in the Attribution.txt files in the cmd
directories.

Also ensure that there is an Attribution.txt file in each of the
cmd directories.

Add this check to make test to fail CI verify jobs if the Attribution.txt files
are missing attributions or executables in cmd are missing an Attribution.txt.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>