Update restAddProfileByYaml response code
Fix #1670

Update the restAddProfileByYaml REST handler to return a '409(Conflict)'
when unable to unmarshal the request body. The reason for this change is
backwards compatibility with EdgeX 1.x. Disregarding backwards
compatibility, the correct response code is '400(Bad Request)'.

Updated unit tests to adhere to the new expectations of the API.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>