Standardize error usage
Fix #1671

Update constructors, error type checks, and tests to use values rather
than references(pointers) for errors.

Update tests to conform to the changes in the production code.

Update error imports to be consistent by aliasing error packages from
different projects or services and not aliasing errors which are defined
within the same project or service.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>