Change service specific env var overrides
Add ability to use underscores within service names, prefixed
to environment variables for overriding a service-specific
configuration item, so that both of the following forms can be
used.

support_notifications_Smtp_Subject
support-notifications_Smtp_Subject

Because the hyphen is not generally permitted by shells as part
of a variable name.

Issue No.: #1519

Signed-off-by: Amit Kumar Chaudhary <a.rudder.free@gmail.com>