SMA GenClients usage not thread-safe
GenClients is an unprotected map updated in multiple locations (as a
direct result of an HTTP request) and referenced elsewhere in the SMA.
Access to it should be protected by a mutex.

Also corrected a panic when registry flag isn't set and config endpoint
is called for an unknown (e.g. not defined in configuration.toml)
service.

Also includes changes related to correcting a race condition in
go-mod-core-contracts (https://github.com/edgexfoundry/go-mod-core-contracts/issues/150).

Fixes https://github.com/edgexfoundry/edgex-go/issues/1763

Signed-off-by: Michael Estrin <m.estrin@dell.com>