Refactor vault client into standalone package with mocks
This is a refactoring of the Vault client originally
found in security-secretstore-setup with re-use and
unit-testability in mind.

The original init and unseal flows have been modified
to remove dependencies on global variables and
hard dependencies on operating system file I/O.
Additonal methods have been added to create Vault
policies and service tokens.

Partially resolves #1677

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>