Launch token provider from security-secretstore-setup.
Replace the placeholder in security-secretstore-setup with
logic to invoke a user-defined process that creates
per-service tokens for the secret-store. At this time
only oneshot providers are supported:
security-secretstore-setup will launch the provider
and wait for it to run to completion.

The new code takes a Context object as input so that
it may be integrated with the new common bootstrap
handler (issue-1751).

Fixes #1798

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>