snapcraft.yaml: set secret store env to false for now
Temporarily set EDGEX_SECURITY_SECRET_STORE to false for edgex-mongo, since it
isn't integrated yet and will fail to run properly until it is, but we want the
default behavior when the env var isn't specified to use security, so we need to
manually shut it off for now.

Also add TODO's to fix this and set the appropriate configuration paths when
the integration is completed.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>