Refactor SecretStoreClient to redude code duplication
This is a refactoring of SecretStoreClient to move request
processing into a common function. This reduces code duplication
and should increase test coverage slightly as a result.

In the future, additional methods will be added to
SecretStoreClient to enable revokation and regeneration of
the root token and stale per-service service tokens.
This refactoring will eanble those methods to be implemented
in less than 10 lines of code in most cases.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>