Uses adopt-info with new 'version' part.
Executing snapcraft sets the snap version for the buiild without
modifying snapcraft.yaml. The version is read from VERSION file
in the root of source tree. The date and git info is appended
to the version as previously. The new version part points at
the snap/local/build-helpers directory, which is unrelated
and is fairly stable. I override-pull and do not execute
'snapcraftctl pull'  in order to prevent these contents from
being moved into parts/version/install, thus resulting in quicker
build and fewer extra files in parts/.

Signed-off-by: Kyle Nitzsche <kyle.nitzsche@canonical.com>