Fix error handling bug in SecretStoreClient.processResponse
The processResponse() internal method takes the response object
and eror object from the previous call to commonRequest().
While processing the error, the code formats an error using
the output error object as a source instead of the input
error object as a source, resulting in a nil access error.
A similar error is made when processing an unexpected response
code: this time a nil error object is always returned.

Unit tests have been added to check regressions.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>