The new boot functionality immediate creates a startup parameter with the parameters '1' and 'internal.BootTimeoutDefault'
BootTimeoutDefault is from the original implementation, in milliseconds, but startup.NewStartupTimer expects seconds.  To fix without breaking the existing usages of internal.BootTimeoutDefault, created a BootTimeoutSecondsDefault and changed all instances of NewStartupTimer to use that.

I also added a BootRetrySecondsDefault for completeness and add it to the calls creating the new startup timer.

Fix #1886

Signed-off-by: Scott C Sosna <me@scottsosna.dev>