Update secret BootstrapHandler to read token file
Fix #1876

Update secret.BootstrapHandler to read the Vault token from a file
specified in the configuration.

Create a new SecretStoreInfo type which encapsulates the SecretConfig
and adds a configuration property for tokenfile information.

Update service specific configurations to use the newly created
SecretStoreInfo rather than SecretConfig so the tokenfile information
can be captured in the struct.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>