Fixes #1814 (nil panic while logging missing res/configuration.toml)
security-secretsstore-setup is not migrating to common bootstrap
in this release. The config module can panic() while calling
deprecated methods to load the configuration while attempting
to log that configuration.toml cannot be found.

This patch sets the global variable before attempting to call
the deprecated configuration loading functions and then forces
the utility to exit if configuration cannot be loaded.
This is a temporary workaround until the utility can be
migrated to use the common bootstrap.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>