Refactor SecretStoreClient to simplfy request/response handling.
SecretStoreClient has three common methods for formatting requests
and processing responses from the secret store:
commonRequest, commonJSONRequest, and commonResponse.

The improved implementation has a common doRequest() method
and takes a parameter struct that enables named parameters to
the function.  This resolves an awkward situation that was raised
in code review that the first lines of commonResponse handle
the errors from commonRequest.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>