Fix security-secretstore-setup secret store initialization bug.
Security-secretstore-setup opens resp-init.json file in
read-only mode prior to checking if secret store
is initialized.  If the secret store must be initialized,
the code is unable to write the init response to disk.

This fix opens resp-init.json in read/write mode initially,
and during the init+unseal flow, resets the file pointer
prior to unsealing.

Fixes #1968

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>