Remove 'LoggingClient' global usage in core-data
Fix #2024

Update core-data to leverage the dependency injection container(DIC)
during bootstrap and pass a 'LoggingClient' instance down the call stack
to eliminate the need for the 'LoggingClient' global variable.

Update tests to use a DIC when loading REST routes as they will need the
DIC to obtain a reference of the 'LoggingClient'. Also, update any
references to methods/functions which have updated signatures to accept
a 'LoggingClient' as part of changes to the production code.

Remove unnecessary test

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>