snapcraft.yaml: add command-chain for setting security secret store env var
If a new config key called security-secret-store is set to off, then we will
set the environment variable EDGEX_SECURITY_SECRET_STORE to false to turn off
secret store (aka Vault currently) integrations in all of the main edgex-go
services and for edgex-mongo.

Note: we don't set that variable here, so this commit specifically will be
broken, but this variable will be setup in a followup commit.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>