snapcraft.yaml: apply patches to openresty during pull stage
If we use the default pull scriptlet, the pull stage doesn't get re-run if parts
OpenResty depends on are changed, so we will just re-run the build stage which
will fail to apply the patches twice. We can instead apply the patches during
the pull stage with a scriptlet, which snapcraft will then clean and re-run for
us if anything changes.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>