snap/many: add app-service-config with rules-engine profile
Add the app-service-configurable service to the snap so that it can be used with
the rules-engine profile in place of the export services for Fuji when a user
wants to use support-rulesengine.

Note: we currently hard-code this to just be the rules-engine profile and a user
is unable to modify it.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>