Remove dbClient global variable from core-data
Fix #2025

Remove dbClient global variable from the core-data service/package(s)
and update the routers to get the dbClient from th DI container and pass
it to the necessary REST handlers. Also, update any functions/methods
which were referring to the dbClient global variable to rather accept an
instance of the DBClient interface.

Update tests to conform to the changes to the production code.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>