Fix orphaned event checksum data for Redis
Fix #2143

Update the deleteEvent function to perform an extra lookup for the
checksum value associated with the event to be deleted and remove the
lingering data. The additional lookup is necessary since there are three
different types of Events(correlation,contract,and redisEvent) being
used within the Redis implementation of the DBClient and the one being
used for common lookups(contract) does not have the checksum information
so it is lost during serialization. The approach is the most direct and
non-invasive solution. Other attempts to clean-up and reduce the
divergence of the types being used resulted in unrelated Blackbox test
failures.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>