Remove chEvent global variable from core-data
Fix #2026

Remove the global variable chEvent and instead use the DI container to
pass the reference to the channel to different parts of the system.

Update function arguments to accept either publishing or consuming Go
channels so that the intent of the function is clear and also to avoid
errors at compile time.

Update tests to work with the new function signatures.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>