Add guard against replacing active PKI
Prevent security-secrets-setup script from terminating
by tailing /dev/null at the end of execution.

This mitigation is needed because blackbox tests
start services one-by-one.  Because dependencies that
are not running are restarted, this causes re-execution
of security-secrets setup, which replaces the active PKI,
which results in fatal TLS verification failures.
Such activity could also be triggered via SMA as well.

In addition, installation of the Vault entrypoint script
is moved up in the secution flow for further robsustness.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>