Remove httpErrorHandler global from coredata
Fix #2030

Remove the httpErrorHandler global variable and instead leverage the DI
container to pass the variable to the necessary functions(router.go).

Update the REST handlers, which were previously referring to the
httpErrorHandler global variable, to accept a httpErrorHandler as
arguments. Update the loadRoutes function to pull the httpErrorHandler
out of the DI container and pass it to the necessary REST handlers.

Update tests to work with the changes to production code.

Clean up imports.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>