Remove Configuration global variable from coredata
Fix # 2031

Remove the 'Configuration' global variable from coredata and instead use
the DI container to pass a reference to which functions/methods require.

Update REST handlers and loadRoutes function to refer to the
configuration provided in the DI container

Update tests to work with the changes to production code.

Fix test which was incorrectly performing a DeepEqual rather than a type
verification check to ensure the factory was returning the correct type
of instance.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>