Enable file-token-provider in docker configuration.
Add configuration to enable the file-token-provider in
the docker configuration.  Also some miscellanous fixes
to connect stdout and stderr of the child process to
the parent's file handles, and changes to the token
configuration file to sync up with existing secret
locations. Make Vault token periodic and indefinitely
renewable with 1h renewal TTL.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>